Refactor entity and DTO relationships and field types
Updated TipoHabitacionDTO to use Long for id fields. Changed HabitacionEntity and TipoHabitacionEntity relationships to use @OneToOne with unique constraints, and removed unnecessary @Column annotations from entities for cleaner code. Adjusted field names for consistency.

+++ Taller2/src/main/java/co/edu/udistrital/taller2/dtos/TipoHabitacionDTO.java
@@ -7,8 +7,8 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class TipoHabitacionDTO {
-    private Int idTipoHabitacion;
-    private Int idHotel;
+    private Long idTipoHabitacion;
+    private Long idHotel;
     private String nombre;
     private Int cantidad;
 }
